add popup feature

+++ front/src/js/list.js
@@ -1,8 +1,11 @@
 import { Toast } from "./toast.js"
-let toast
+import { EditTaskPopup } from "./edit-task.popup.js"
+
+let toast, editTaskPopup
 
 document.addEventListener("DOMContentLoaded", function() {
     toast = new Toast()
+    editTaskPopup = new EditTaskPopup()
     getTasksList({ urgency__in: "1" })
         .then(tasks => {
             renderList(document.getElementById("tasks-list-1"), tasks)
@@ -18,18 +21,61 @@ document.addEventListener("DOMContentLoaded", function() {
 })
 
 window.removeTask = function removeTask(id) {
-    axios.delete(`http://localhost:3000/tasks/${id}`)
-    .then(() => {
-        document.getElementById(`task-${id}`).remove()
-        toast.show(`Task "${id}" has been successfully deleted.`, "success")
-    })
-    .catch(error => {
-        console.log(error)
-        toast.show(`Failed to delete task "${id}".`, "error")
-    })
+    let res = confirm(`Do you really want to remove the ${id}?`)
+    if (res) {
+        axios.delete(`http://localhost:3000/tasks/${id}`)
+            .then(() => {
+                removeTaskFromDOM(id)
+                toast.show(`Task "${id}" has been successfully deleted.`, "success")
+            })
+            .catch(error => {
+                console.log(error)
+                toast.show(`Failed to delete task "${id}".`, "error")
+            })
+    }
 }
+window.editTask = function editTask(id) {
+    editTaskPopup.show(id)
+    let timer = setInterval(() => {
+        if (editTaskPopup.isClosed()) {
+            clearInterval(timer)
+            if(editTaskPopup.updatedTask) {
+                console.log("editTaskPopup.updatedTask", editTaskPopup.updatedTask)
+                let filters, tasksListId
 
-function getTasksList(filters) {
+                switch (editTaskPopup.updatedTask.urgency){
+                    case 1:
+                        filters = { urgency__in: "1" }
+                        tasksListId = "tasks-list-1"
+                        break
+                    case 2:
+                    case 3:
+                        filters = { urgency__in: "2,3" }
+                        tasksListId = "tasks-list-2"
+                        break
+                    case 4:
+                    case 5:
+                        filters = { urgency__in: "4,5" }
+                        tasksListId = "tasks-list-3"
+                        break
+                }
+                getTasksList(filters)
+                    .then(tasks => {
+                        document.getElementById(tasksListId).innerHTML = ""
+                        renderList(document.getElementById(tasksListId), tasks)
+                    })
+                toast.show(`Task updated`, `success`)
+            }
+            console.log("Popup Close")
+        }
+    }, 500)
+
+}
+    function removeTaskFromDOM(id) {
+        document.getElementById(`task-${id}`).remove()
+    }
+
+    function getTasksList(filters) {
     return axios.get('http://localhost:3000/tasks', {
         params: { 
             sort: "-updatedAt",
@@ -67,9 +113,10 @@ function renderList(rootElement, tasks) {
                                     </button>
                                 </div>
                                 <div class="task-action">
-                                    <button>
+                                    <button onclick=editTask("${task._id}")>
                                         <a href="#">
-                                            <img src="https://img.icons8.com/metro/26/000000/pencil.png"/>                                                </a>
+                                            <img src="https://img.icons8.com/metro/26/000000/pencil.png"/>  
+                                        </a>
                                     </button>
                                 </div>` 
         taskElement.append(taskActions)
